+++ src/commands/SetTeam/NoTeamsFoundError.ts
@@ -1,5 +1,5 @@
 export class NoTeamsFoundError extends Error {
-	constructor() {
-		super("No teams found");
-	}
+  constructor() {
+    super('No teams found')
+  }
 }