+++ src/logger.ts
@@ -1,8 +1,7 @@
-import pino from 'pino';
-import { window } from 'vscode';
-const channelName = "taskstarter";
-const outputChannel = window.createOutputChannel(channelName);
-const logger = pino({ level: "debug", });
+import pino from 'pino'
+// import { window } from 'vscode'
+// const channelName = 'taskstarter'
+// const outputChannel = window.createOutputChannel(channelName)
+const logger = pino({ level: 'debug' })
 
-
-export { logger };
+export { logger }